feat: start currency api
