+++ src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/data/stored/CurrencyStorage.kt
@@ -1,0 +1,7 @@
+package tech.thatgravyboat.skyblockapi.api.data.stored
+
+internal object CurrencyStorage {
+
+
+
+}
